Moved some packages to `internal` directory
