+++ internal/database/err.go
