+++ params/strings.go
@@ -31,10 +31,9 @@ const (
 	listDesc        = "Показать только список групп"
 	icsDesc         = "Вывод в ics файл"
 	helpDesc        = "Показать эту помощь"
-	proxyDesc       = "Использовать прокси:\n" +
-		"  <протокол://адрес:порт>"
-	sessionDesc = "Расписание сессии (игнорирует выбор недели: --week, --next, --current)"
-	clearDesc   = "Очистить кэш групп"
-	workdirDesc = "Путь рабочей директории (кэш) (по умолчанию равен pwd)"
-	outPathDesc = "Путь для вывода (если не задан то равен --workdir)"
+	proxyDesc       = "Использовать прокси: <протокол://адрес:порт>"
+	sessionDesc     = "Расписание сессии (игнорирует выбор недели: --week, --next, --current)"
+	clearDesc       = "Очистить кэш групп"
+	workdirDesc     = "Путь рабочей директории (кэш) (по умолчанию равен pwd)"
+	outPathDesc     = "Путь для вывода (если не задан то равен --workdir)"
 )
